requirements.txt
streamlit
google-generativeai
python-dotenv
mysql-connector-python #As am using mysql workbench which is loacally installed